requirements.txt
akshare==1.16.99
lightgbm==4.0.0
optuna==3.4.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
shap==0.44.0
imblearn==0.12.0
matplotlib==3.8.2
seaborn==0.13.0
statsmodels==0.14.0
tqdm==4.66.0
pyyaml==6.0.1